pyproject.toml
[project]
name = "otel-api-scraper"
version = "0.1.0"
description = "Config-driven async API scraper emitting OpenTelemetry metrics and logs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx[http2]==0.28.1",
    "pyyaml>=6.0.2",
    "pydantic>=2.9,<3.0",
    "python-dotenv>=1.0.1",
    "opentelemetry-api==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-exporter-otlp-proto-http==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "apscheduler[redis]==3.11.1",
    "aiosqlite==0.21.0",
    "valkey>=6.0.0",
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "python-dateutil>=2.9.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=3.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
otel-api-scraper = "otel_api_scraper.runner:main"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version = "0.1.0"
version_files = [
    "pyproject.toml:version"
]
major_version_zero = true
check_consistency = true
use_shortcuts = true

[tool.commitizen.customize]
message_template = "{{change_type}}{% if scope %}({{scope}}){% endif %}: {{message}}"
example = "feat(parser): add ability to parse arrays"
schema = """
<type>(<scope>): <subject>

<body>

<footer>
"""
schema_pattern = "^(feat|fix|docs|style|refactor|perf|test|build|ci|chore|revert)(\\(.+\\))?: .+$"
change_type_order = ["BREAKING CHANGE", "feat", "fix", "refactor", "perf"]
commit_parser = "^(?P<change_type>feat|fix|docs|style|refactor|perf|test|build|ci|chore|revert)(?:\\((?P<scope>[^()\r\n]*)\\)|\\()?(?P<breaking>!)?:\\s(?P<message>.*)?"
changelog_pattern = "^(feat|fix|docs|style|refactor|perf|test|build|ci|chore|revert)?(\\(.+\\))?(!)?"

[[tool.commitizen.customize.change_type_map]]
change_type = "feat"
changelog = "Features"

[[tool.commitizen.customize.change_type_map]]
change_type = "fix"
changelog = "Bug Fixes"

[[tool.commitizen.customize.change_type_map]]
change_type = "refactor"
changelog = "Code Refactoring"

[[tool.commitizen.customize.change_type_map]]
change_type = "perf"
changelog = "Performance Improvements"
